+++ Source/Events/EventHorizon/IExternalEvent.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Dolittle.Events
+namespace Dolittle.Events.EventHorizon
 {
     /// <summary>
     /// Defines the basics of an external event.
